requirements.txt
Django==3.2.8
django-templated-mail==1.1.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
psycopg2-binary==2.9.1
pycparser==2.20
six==1.16.0

